refactor: TCD-3944 - 소스 정리
- 위젯 분리
