+++ lib/screen/widget/my_google_map.dart
@@ -1,19 +1,18 @@
 import 'package:flutter/material.dart';
-import 'package:geolocator/geolocator.dart';
 import 'package:go_to_work/constant/constants.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 
 class MyGoogleMap extends StatefulWidget {
   final LatLng baseLntLng;
   final MapCreatedCallback onMapCreated;
-  final ValueChanged<bool> withinRanged;
+  final bool isWithin;
   final bool isDone;
 
   const MyGoogleMap({
     super.key,
     required this.baseLntLng,
     required this.onMapCreated,
-    required this.withinRanged,
+    required this.isWithin,
     required this.isDone,
   });
 
@@ -22,6 +21,8 @@ class MyGoogleMap extends StatefulWidget {
 }
 
 class _MyGoogleMapState extends State<MyGoogleMap> {
+  GoogleMapController? googleMapController;
+
   late Marker baseMarker;
   late Color baseCircleColor;
 
@@ -37,47 +38,28 @@ class _MyGoogleMapState extends State<MyGoogleMap> {
 
   @override
   Widget build(BuildContext context) {
-    return StreamBuilder(
-      stream: Geolocator.getPositionStream(),
-      builder: (context, snapshot) {
-        bool withinRanged = false;
-        if (snapshot.hasData) {
-          withinRanged = checkLocation(snapshot.data!);
-          widget.withinRanged(withinRanged);
-        }
-
-        return GoogleMap(
-          initialCameraPosition: CameraPosition(
-            target: widget.baseLntLng,
-            zoom: 17,
-          ),
-          mapType: MapType.normal,
-          myLocationEnabled: true,
-          myLocationButtonEnabled: false,
-          circles: {
-            baseCircle(done: widget.isDone, withinRanged: withinRanged),
-          },
-          markers: {
-            baseMarker,
-          },
-          onMapCreated: widget.onMapCreated,
-        );
+    return GoogleMap(
+      initialCameraPosition: CameraPosition(
+        target: widget.baseLntLng,
+        zoom: 17,
+      ),
+      mapType: MapType.normal,
+      myLocationEnabled: true,
+      myLocationButtonEnabled: false,
+      circles: {
+        baseCircle(done: widget.isDone, withinRanged: widget.isWithin),
       },
+      markers: {
+        baseMarker,
+      },
+      onMapCreated: onGoogleMapCreated,
     );
   }
 
-  bool checkLocation(Position position) {
-    final start = position;
-    final end = widget.baseLntLng;
-
-    final distance = Geolocator.distanceBetween(
-      start.latitude,
-      start.longitude,
-      end.latitude,
-      end.longitude,
-    );
-
-    return distance <= MyGoogleMapConstants.OK_DISTANCE;
+  @override
+  void dispose() {
+    googleMapController?.dispose();
+    super.dispose();
   }
 
   Circle baseCircle({
@@ -104,5 +86,10 @@ class _MyGoogleMapState extends State<MyGoogleMap> {
       strokeColor: color,
       strokeWidth: 1,
     );
+  }
+
+  void onGoogleMapCreated(GoogleMapController controller) {
+    googleMapController = controller;
+    widget.onMapCreated(controller);
   }
 }
